+++ source/components/player/states/concurrent_states/FastAttackAirActionState.h
@@ -17,7 +17,7 @@ class FastAttackAirActionState : public AirborneActionState {
 	float interruptibleTime;
 
 	CHandle hitboxHandle;
-	int damage = 1;
+	float damage = 1.f;
 	AttackPhases phase = AttackPhases::Launch;
 
 	float powerToGet = 1000.f;
